Consolidating loss code
