+++ src/MLP.py
@@ -4,8 +4,7 @@ reference: https://medium.com/biaslyai/pytorch-introduction-to-neural-network-fe
 
 import torch
 from src.Handler import *
-import numpy as np
-from math import sqrt
+import src.CustomLoss as CustomLoss
 
 
 class MLP(torch.nn.Module):
@@ -41,11 +40,11 @@ class MLPHandler(Handler):
 
     def train(self, x, y):
         if self.loss_method == 'MSE':
-            criterion = mse_loss
+            criterion = CustomLoss.mse_loss
         elif self.loss_method == 'Returns':
-            criterion = return_loss
+            criterion = CustomLoss.return_loss
         elif self.loss_method == 'Sharpe':
-            criterion = sharpe_loss
+            criterion = CustomLoss.sharpe_loss
         else:
             raise Exception("Invalid loss method")
         optimizer = torch.optim.Adam(self.model.parameters(), lr=self.learning_rate)
@@ -84,27 +83,3 @@ class MLPHandler(Handler):
         y_pred = self.model(x)
         loss = criterion(y_pred.squeeze(), y)
         return loss, y_pred
-
-def return_loss(inputs, target):
-    volatility_scaling = 1
-    sig_tgt = .15
-    return torch.mean(np.sign(target) * inputs) * -1
-
-def mse_loss(output, target):
-    volatility_scaling = 1
-    loss = torch.mean((output - (target / volatility_scaling))**2)
-    return loss
-
-def binary_loss(inputs, target):
-    volatility_scaling = 1
-    loss = torch.log(target)
-
-def sharpe_loss(inputs, target):
-    n_days = 252
-    # R_it is the return given by the targets
-    R_it = torch.sum(target ** 2) / len(inputs)
-
-    loss = torch.mean(inputs) * sqrt(n_days)
-    loss /= torch.sqrt(torch.abs(R_it - torch.mean(torch.pow(inputs, 2))))
-
-    return loss * -1
